README.md
In this project we have used data visualization. Data visualization is the representation of the data in a graphical form. It converts the raw data present into a much more representable form, as a result this data can be communicated to the user more efficiently. Data Visualization makes the data more understandable and user-friendly. In the age of big data and data analytics the incomprehensible amount of data present can be put to a great use with the help of data visualization.